changed professor.js to TS
